Refactor CustomTimer to show time in milliseconds for storage and leaderboard purposes. Edited Storage to take account for the milliseconds Edited handleWin in GamePage to fix a bug where the timer label lags or does not match the win time

+++ src/main/java/minesweeper/Storage.java
@@ -144,16 +144,16 @@ public class Storage {
      * Load time from time.txt file.
      * Time is an integer value (represents seconds).
      *
-     * @return Time (integer value).
+     * @return Time (long value for milliseconds).
      * @throws StorageException
      */
-    public int loadTime() throws StorageException {
+    public long loadTime() throws StorageException {
         try {
             List<String> lines = Files.readAllLines(this.timeFilePath);
             if (lines.isEmpty()) {
                 throw new StorageException("No Existing Timing");
             }
-            return Integer.parseInt(lines.getFirst().trim());
+            return Long.parseLong(lines.getFirst().trim());
         } catch (IOException loadError) {
             throw new StorageException("No Existing Timing");
         }
